+++ .github/helpers/contracts/dummy-dao/constans.js
@@ -1,0 +1,16 @@
+const TOKENLIST = {
+  rinkeby: [
+    {
+      name: 'Dai',
+      address: '0xc7AD46e0b8a400Bb3C915120d284AafbA8fc4735',
+    },
+    {
+      name: 'Uni',
+      address: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984',
+    },
+  ],
+  'arbitrum-rinkeby': [],
+  mumbai: [],
+};
+
+module.exports = {TOKENLIST};